+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 # main
 
+* Remove cop `Rails/TableName` in favor of [Rails/TableNameAssignment](https://docs.rubocop.org/rubocop-rails/cops_rails.html#railstablenameassignment)
+  and enable it in base config (Fix [#45](https://github.com/petalmd/rubocop-petal/issues/45)). ([#48](https://github.com/petalmd/rubocop-petal/pull/48))
 * Added cop `Migration/ChangeTableReferences` ([#47](https://github.com/petalmd/rubocop-petal/pull/47))
 * Added cop `Migration/AlwaysBulkChangeTable` ([#46](https://github.com/petalmd/rubocop-petal/pull/46))